apps/snapodds/src/services/auth/auth-interceptor.service.ts
import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
import { Injectable } from '@angular/core';
import { Observable } from 'rxjs';
import { AuthService } from './auth.service';

@Injectable()
export class AuthHttpInterceptor implements HttpInterceptor {
  constructor(private authService: AuthService) {}

  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
    if (this.authService.hasValidAccessToken()) {
      request = request.clone({
        setHeaders: { Authorization: `Bearer ${this.authService.getAccessToken()}` },
      });
    }

    return next.handle(request);
  }
}
